# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/insert": {
      "post": {
        "tags": [
          "Station"
        ],
        "description": "Endpoint to insert a station",
        "parameters": [
          {
            "name": "newStation",
            "in": "body",
            "description": "Information about the station",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Station"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "insered station.",
            "schema": {
              "$ref": "#/definitions/Station"
            }
          }
        }
      }
    },
    "/getNearStaions/{longitude}/{latitude}": {
      "get": {
        "tags": [
          "Station"
        ],
        "description": "Endpoint to get all stations in a round of 5000m to the center",
        "parameters": [
          {
            "name": "longitude",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "center longitude"
          },
          {
            "name": "latitude",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "center latitude"
          }
        ],
        "responses": {
          "200": {
            "description": "near station.",
            "schema": {
              "$ref": "#/definitions/Station"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Station": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Central Park"
        },
        "location": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "Point"
            },
            "coordinates": {
              "type": "array",
              "example": [
                -73.97,
                40.77
              ],
              "items": {
                "type": "number"
              }
            }
          }
        }
      },
      "required": [
        "name",
        "location"
      ]
    }
  }
}